requirements/heroku.txt
-r _base.txt
PyTumblr==0.0.5
boto==2.8.0
django-storages==1.1.8
gunicorn==17.5
httplib2==0.8
oauth2==1.5.211
python-dateutil==1.5
